requirements.txt
Flask==2.1.0
gunicorn==20.1.0
google-cloud-storage~=1.44.0
tensorflow-cpu~=2.8.0